+++ tests/main.cpp
@@ -6,6 +6,7 @@
 #include "json.h"
 #include "format.h"
 #include "hash-map.h"
+#include "byte-array.h"
 #include "d-bus.h"
 
 void test_with_log(const char *name, void(*test_func)())
@@ -94,6 +95,15 @@ int main(int argc, char *argv[])
     test_with_log("hash_map_insert", tests::hash_map_insert);
 
     test_with_log("hash_map_get", tests::hash_map_get);
+
+    //============
+    // ByteArray
+    //============
+    test_with_log("byte_array_init", tests::byte_array_init);
+
+    test_with_log("byte_array_equal", tests::byte_array_equal);
+
+    test_with_log("byte_array_subset", tests::byte_array_subset);
 
     //=========
     // DBus